[test] 댓글 작성 삭제 테스트 작성
댓글을 작성하고 삭제하는 테스트를 작성하였습니다.

+++ src/main/java/lass9436/comment/servlet/CommentServlet.java
@@ -46,6 +46,21 @@ public class CommentServlet extends HttpServlet {
 	@Override
 	protected void doDelete(HttpServletRequest req, HttpServletResponse resp) {
 		long seq = Long.parseLong(req.getParameter("seq"));
+		Comment comment = commentRepository.findByCommentSeq(seq);
+
+		if (comment == null) {
+			resp.setStatus(HttpServletResponse.SC_NOT_FOUND);
+			return;
+		}
+		if (req.getSession(false) == null){
+			resp.setStatus(HttpServletResponse.SC_UNAUTHORIZED);
+			return;
+		}
+		if (comment.getUserSeq() != (long)req.getSession(false).getAttribute("userSeq")){
+			resp.setStatus(HttpServletResponse.SC_FORBIDDEN);
+			return;
+		}
+
 		commentRepository.deleteByCommentSeq(seq);
 		resp.setStatus(HttpServletResponse.SC_OK);
 	}
